Update main.py 5/17/17 1:50pm
updated TorqueDown() to have a check in case the system stalls before pressing down on the rails in preparation for integrating a new design proposal:
Have the Rail Actuators approach the rails at a low EMF so they stall at low amounts. Then increase the EMF so they press down on the
rails hard, hopefully at a distance so small that nothing important gets pinched.