Update main.py 5/20/17 3:30 pm to 5/21/17 2:20 am
Program edited in multiple spots.
Function ImportBoltPattern() has been updated to not accept "B" as a valid input
Function TorqueDown() was completely rewritten. Changes were made to allow the program to user the rail actuators to approach the rails at a low torque. If an stall occurs, it homes and retries. When the rail actuators have finished their approach, the torque increases and presses down until stall. If it does not stall and reaches position above granite, it stops, homes, and retries from the top. If it stalled, torque bolts down and then home rail actuators and screw drivers.